requirements.txt
# Core Flask
Flask==3.0.0
Werkzeug==3.0.1

# NLP & Chatbot
nltk==3.8.1

# API Requests
requests==2.31.0

# Password Hashing
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# Database (if you want to upgrade from SQLite later)
# Flask-SQLAlchemy==3.1.1

# Input Sanitization
bleach==6.1.0

# JSON handling (built-in, but listing for clarity)
# json - built-in

# Additional utilities
python-dateutil==2.8.2
